move node invocations to sys
