+++ cli/src/commands/_shared/authed-command.ts
@@ -2,7 +2,7 @@ import type { KyInstance } from 'ky'
 import type { HostsBundle } from '../../auth/hosts.js'
 import type { AppInfoCache } from '../../cache/app-info.js'
 import type { Command } from '../../framework/command.js'
-import type { IOStreams } from '../../io/streams.js'
+import type { IOStreams } from '../../sys/io/streams'
 import { META_PROBE_TIMEOUT_MS, MetaClient } from '../../api/meta.js'
 import { loadHosts } from '../../auth/hosts.js'
 import { loadAppInfoCache } from '../../cache/app-info.js'
@@ -10,8 +10,9 @@ import { loadNudgeStore } from '../../cache/nudge-store.js'
 import { BaseError } from '../../errors/base.js'
 import { ErrorCode } from '../../errors/codes.js'
 import { formatErrorForCli } from '../../errors/format.js'
+import { getEnv } from '../../env/registry.js'
 import { createClient } from '../../http/client.js'
-import { realStreams } from '../../io/streams.js'
+import { realStreams } from '../../sys/io/streams'
 import { resolveConfigDir } from '../../store/dir.js'
 import { hostWithScheme } from '../../util/host.js'
 import { versionInfo } from '../../version/info.js'
@@ -38,6 +39,7 @@ export async function buildAuthedContext(
   opts: AuthedContextOptions,
 ): Promise<AuthedContext> {
   const configDir = resolveConfigDir()
+  const io = realStreams(opts.format ?? '')
   const bundle = await loadHosts(configDir)
   if (bundle === undefined || bundle.tokens?.bearer === undefined || bundle.tokens.bearer === '') {
     const err = new BaseError({
@@ -45,16 +47,15 @@ export async function buildAuthedContext(
       message: 'not logged in',
       hint: 'run \'difyctl auth login\'',
     })
-    cmd.error(formatErrorForCli(err, { format: opts.format, isErrTTY: process.stderr.isTTY }), { exit: err.exit() })
+    cmd.error(formatErrorForCli(err, { format: opts.format, isErrTTY: io.isErrTTY }), { exit: err.exit() })
   }
 
   const host = hostWithScheme(bundle.current_host, bundle.scheme)
   const retryAttempts = resolveRetryAttempts({
     flag: opts.retryFlag,
-    env: (k: string) => process.env[k],
+    env: getEnv,
   })
   const http = createClient({ host, bearer: bundle.tokens.bearer, retryAttempts })
-  const io = realStreams(opts.format ?? '')
 
   const cache = opts.withCache === true ? await loadAppInfoCache({ configDir }) : undefined
