+++ cli/src/commands/auth/devices/_shared/devices.test.ts
@@ -8,7 +8,7 @@ import { afterEach, beforeEach, describe, expect, it } from 'vitest'
 import { startMock } from '../../../../../test/fixtures/dify-mock/server.js'
 import { saveHosts } from '../../../../auth/hosts.js'
 import { createClient } from '../../../../http/client.js'
-import { bufferStreams } from '../../../../io/streams.js'
+import { bufferStreams } from '../../../../sys/io/streams'
 import { runDevicesList, runDevicesRevoke } from './devices.js'
 
 class MemStore implements TokenStore {
