+++ cli/src/commands/auth/devices/_shared/devices.ts
@@ -2,15 +2,15 @@ import type { SessionRow } from '@dify/contracts/api/openapi/types.gen'
 import type { KyInstance } from 'ky'
 import type { HostsBundle } from '../../../../auth/hosts.js'
 import type { TokenStore } from '../../../../auth/store.js'
-import type { IOStreams } from '../../../../io/streams.js'
+import type { IOStreams } from '../../../../sys/io/streams'
 import { unlink } from 'node:fs/promises'
 import { join } from 'node:path'
 import { AccountSessionsClient } from '../../../../api/account-sessions.js'
 import { HOSTS_FILE_NAME } from '../../../../auth/hosts.js'
 import { BaseError } from '../../../../errors/base.js'
 import { ErrorCode } from '../../../../errors/codes.js'
-import { colorEnabled, colorScheme } from '../../../../io/color.js'
-import { runWithSpinner } from '../../../../io/spinner.js'
+import { colorEnabled, colorScheme } from '../../../../sys/io/color.js'
+import { runWithSpinner } from '../../../../sys/io/spinner.js'
 
 export type DevicesListOptions = {
   readonly io: IOStreams
