+++ cli/src/commands/auth/login/login.ts
@@ -1,7 +1,7 @@
 import type { CodeResponse, PollSuccess } from '../../../api/oauth-device.js'
 import type { HostsBundle, StorageMode, Workspace } from '../../../auth/hosts.js'
 import type { TokenStore } from '../../../auth/store.js'
-import type { IOStreams } from '../../../io/streams.js'
+import type { IOStreams } from '../../../sys/io/streams'
 import type { BrowserEnv, BrowserOpener } from '../../../util/browser.js'
 import type { Clock } from './device-flow.js'
 import * as os from 'node:os'
@@ -10,7 +10,7 @@ import { DeviceFlowApi } from '../../../api/oauth-device.js'
 import { saveHosts } from '../../../auth/hosts.js'
 import { selectStore } from '../../../auth/store.js'
 import { createClient } from '../../../http/client.js'
-import { colorEnabled, colorScheme } from '../../../io/color.js'
+import { colorEnabled, colorScheme } from '../../../sys/io/color.js'
 import { decideOpen, OpenDecision, openUrl, realEnv } from '../../../util/browser.js'
 import { bareHost, DEFAULT_HOST, resolveHost, validateVerificationURI } from '../../../util/host.js'
 import { awaitAuthorization, realClock } from './device-flow.js'
