+++ cli/src/commands/auth/logout/index.ts
@@ -2,8 +2,8 @@ import type { KyInstance } from 'ky'
 import { loadHosts } from '../../../auth/hosts.js'
 import { selectStore } from '../../../auth/store.js'
 import { createClient } from '../../../http/client.js'
-import { runWithSpinner } from '../../../io/spinner.js'
-import { realStreams } from '../../../io/streams.js'
+import { runWithSpinner } from '../../../sys/io/spinner.js'
+import { realStreams } from '../../../sys/io/streams'
 import { resolveConfigDir } from '../../../store/dir.js'
 import { hostWithScheme } from '../../../util/host.js'
 import { DifyCommand } from '../../_shared/dify-command.js'
