+++ cli/src/commands/auth/whoami/index.ts
@@ -1,6 +1,6 @@
 import { loadHosts } from '../../../auth/hosts.js'
 import { Flags } from '../../../framework/flags.js'
-import { realStreams } from '../../../io/streams.js'
+import { realStreams } from '../../../sys/io/streams'
 import { resolveConfigDir } from '../../../store/dir.js'
 import { DifyCommand } from '../../_shared/dify-command.js'
 import { runWhoami } from './whoami.js'
