+++ cli/src/commands/auth/whoami/whoami.ts
@@ -1,5 +1,5 @@
 import type { HostsBundle } from '../../../auth/hosts.js'
-import type { IOStreams } from '../../../io/streams.js'
+import type { IOStreams } from '../../../sys/io/streams'
 import { BaseError } from '../../../errors/base.js'
 import { ErrorCode } from '../../../errors/codes.js'
