+++ cli/src/commands/describe/app/run.ts
@@ -1,12 +1,13 @@
 import type { KyInstance } from 'ky'
 import type { HostsBundle } from '../../../auth/hosts.js'
 import type { AppInfoCache } from '../../../cache/app-info.js'
-import type { IOStreams } from '../../../io/streams.js'
+import type { IOStreams } from '../../../sys/io/streams'
 import { AppMetaClient } from '../../../api/app-meta.js'
 import { AppsClient } from '../../../api/apps.js'
-import { runWithSpinner } from '../../../io/spinner.js'
-import { nullStreams } from '../../../io/streams.js'
+import { runWithSpinner } from '../../../sys/io/spinner.js'
+import { nullStreams } from '../../../sys/io/streams'
 import { FieldInfo, FieldInputSchema, FieldParameters } from '../../../types/app-meta.js'
+import { getEnv } from '../../../sys/index.js'
 import { resolveWorkspaceId } from '../../../workspace/resolver.js'
 import { AppDescribeOutput } from './handlers.js'
 
@@ -27,7 +28,7 @@ export type DescribeAppDeps = {
 }
 
 export async function runDescribeApp(opts: DescribeAppOptions, deps: DescribeAppDeps): Promise<AppDescribeOutput> {
-  const env = deps.envLookup ?? ((k: string) => process.env[k])
+  const env = deps.envLookup ?? getEnv
   const wsId = resolveWorkspaceId({ flag: opts.workspace, env: env('DIFY_WORKSPACE_ID'), bundle: deps.bundle })
   const apps = new AppsClient(deps.http)
   const meta = new AppMetaClient({ apps, host: deps.host, cache: deps.cache })
