+++ cli/src/commands/get/app/run.ts
@@ -1,12 +1,13 @@
 import type { AppDescribeResponse, AppListResponse, AppMode } from '@dify/contracts/api/openapi/types.gen'
 import type { KyInstance } from 'ky'
 import type { HostsBundle } from '../../../auth/hosts.js'
-import type { IOStreams } from '../../../io/streams.js'
+import type { IOStreams } from '../../../sys/io/streams'
 import { AppsClient } from '../../../api/apps.js'
 import { WorkspacesClient } from '../../../api/workspaces.js'
-import { runWithSpinner } from '../../../io/spinner.js'
-import { nullStreams } from '../../../io/streams.js'
+import { runWithSpinner } from '../../../sys/io/spinner.js'
+import { nullStreams } from '../../../sys/io/streams'
 import { LIMIT_DEFAULT, parseLimit } from '../../../limit/limit.js'
+import { getEnv } from '../../../sys/index.js'
 import { resolveWorkspaceId } from '../../../workspace/resolver.js'
 import { AppListOutput, AppRow } from './handlers.js'
 
@@ -38,7 +39,7 @@ export type GetAppResult = {
 }
 
 export async function runGetApp(opts: GetAppOptions, deps: GetAppDeps): Promise<GetAppResult> {
-  const env = deps.envLookup ?? ((k: string) => process.env[k])
+  const env = deps.envLookup ?? getEnv
   const appsFactory = deps.appsFactory ?? ((h: KyInstance) => new AppsClient(h))
   const wsFactory = deps.workspacesFactory ?? ((h: KyInstance) => new WorkspacesClient(h))
