+++ cli/src/commands/run/app/_strategies/streaming-structured.ts
@@ -1,9 +1,9 @@
 import type { SseEvent } from '../../../../http/sse.js'
 import type { RunContext, RunStrategy } from './index.js'
 import { buildRunBody } from '../../../../api/app-run.js'
-import { colorEnabled, colorScheme } from '../../../../io/color.js'
-import { startSpinner } from '../../../../io/spinner.js'
-import { extractThinkBlocks, stripThinkBlocks } from '../../../../io/think-filter.js'
+import { colorEnabled, colorScheme } from '../../../../sys/io/color.js'
+import { startSpinner } from '../../../../sys/io/spinner.js'
+import { extractThinkBlocks, stripThinkBlocks } from '../../../../sys/io/think-filter.js'
 import { chatConversationHint, newAppRunObject, RUN_MODES } from '../handlers.js'
 import { renderHitlHint, renderHitlOutput } from '../hitl-render.js'
 import { collect, HitlPauseError } from '../sse-collector.js'
