+++ cli/src/commands/run/app/stream-handlers.ts
@@ -3,8 +3,8 @@ import type { StreamPrinter } from '../../../printers/stream-printer.js'
 import type { HitlPausePayload } from './sse-collector.js'
 import { newError } from '../../../errors/base.js'
 import { ErrorCode } from '../../../errors/codes.js'
-import { colorEnabled, colorScheme } from '../../../io/color.js'
-import { ThinkChunkFilter } from '../../../io/think-filter.js'
+import { colorEnabled, colorScheme } from '../../../sys/io/color.js'
+import { ThinkChunkFilter } from '../../../sys/io/think-filter.js'
 import { RUN_MODES } from './handlers.js'
 import { HitlPauseError } from './sse-collector.js'
