+++ cli/src/sys/io/color.ts
