+++ cli/src/sys/io/think-filter.test.ts
