+++ cli/src/version/render.ts
@@ -1,5 +1,5 @@
 import type { VersionReport } from './probe.js'
-import { colorScheme } from '../io/color.js'
+import { colorScheme } from '../sys/io/color.js'
 
 const RC_WARNING_LINES = [
   'WARNING: This build is a release candidate. It is in beta test, not stable,',
